Implement owner validation for CRUD operations and improve error handling

+++ src/controllers/video.controller.js
@@ -8,7 +8,7 @@ import { ApiResponse } from "../utils/ApiResponse.js"
 const getAllVideos = asyncHandler(async (req, res) => {
     const { page = 1, limit = 10, query, sortBy, sortType, userId } = req.query
     //TODO: get all videos based on query, sort, pagination
-    const videos = await VideoModel.findOne({});
+    const videos = await VideoModel.find({ owner: req.user?._id });
 
     return res.status(200).json(new ApiResponse(200, videos, 'All videos'))
 })
@@ -65,7 +65,7 @@ const updateVideo = asyncHandler(async (req, res) => {
         throw new ApiError(500, 'Something went wrong while uploading avatar')
     }
 
-    const video = await VideoModel.findByIdAndUpdate(videoId, {
+    const video = await VideoModel.findByIdAndUpdate({ _id: videoId, owner: req.user._id }, {
         $set: {
             title,
             description,
@@ -80,7 +80,7 @@ const deleteVideo = asyncHandler(async (req, res) => {
     const { videoId } = req.params
     //TODO: delete video
 
-    await VideoModel.findByIdAndDelete(videoId);
+    await VideoModel.findByIdAndDelete({ _id: videoId, owner: req.user._id });
 
     return res.status(200).json(new ApiResponse(200, {}, 'Video deleted'))
 })
@@ -88,7 +88,7 @@ const deleteVideo = asyncHandler(async (req, res) => {
 const togglePublishStatus = asyncHandler(async (req, res) => {
     const { videoId } = req.params
     const { status } = req.body;
-    const video = await VideoModel.findByIdAndUpdate(videoId, {
+    const video = await VideoModel.findByIdAndUpdate({ _id: videoId, owner: req.user._id }, {
         $set: {
             isPublished: status
         }
